✨ feat: split resolver.tx and SpecRenderer.tsx
