+++ starters/react-router/app/integration/resolvers/SpecRenderer.tsx
@@ -1,0 +1,16 @@
+import type { ReactNode } from "react";
+import type { Spec } from "@json-render/core";
+import { JSONUIProvider, Renderer } from "@json-render/react";
+import { registry } from "~/integration/resolvers/registry";
+
+interface SpecRendererProps {
+  spec: Spec;
+}
+
+export function SpecRenderer({ spec }: SpecRendererProps): ReactNode {
+  return (
+    <JSONUIProvider registry={registry}>
+      <Renderer spec={spec} registry={registry} />
+    </JSONUIProvider>
+  );
+}
